+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/types/resources/HealPower.java
@@ -1,6 +1,7 @@
 package com.robertx22.age_of_exile.database.data.stats.types.resources;
 
 import com.robertx22.age_of_exile.database.data.stats.Stat;
+import com.robertx22.age_of_exile.database.data.stats.StatScaling;
 import com.robertx22.age_of_exile.database.data.stats.effects.resource.IncreaseHealingEffect;
 import com.robertx22.age_of_exile.uncommon.enumclasses.Elements;
 import com.robertx22.age_of_exile.uncommon.interfaces.IStatEffect;
@@ -12,6 +13,7 @@ public class HealPower extends Stat implements IStatEffects {
     private HealPower() {
         this.add$To$toTooltip = false;
         this.statGroup = StatGroup.RESTORATION;
+        this.scaling = StatScaling.SLOW;
     }
 
     public static HealPower getInstance() {