+++ src/main/java/com/robertx22/age_of_exile/mmorpg/registers/common/items/InscribingItemRegister.java
@@ -1,5 +1,6 @@
 package com.robertx22.age_of_exile.mmorpg.registers.common.items;
 
+import com.robertx22.age_of_exile.player_skills.items.fishing.ScribeInkItem;
 import com.robertx22.age_of_exile.player_skills.items.foods.SkillItemTier;
 import com.robertx22.age_of_exile.player_skills.items.protection_tablets.BlankTabletItem;
 import com.robertx22.age_of_exile.player_skills.items.protection_tablets.BlankTabletTier;
@@ -7,9 +8,12 @@ import com.robertx22.age_of_exile.player_skills.items.protection_tablets.Protect
 import com.robertx22.age_of_exile.player_skills.items.protection_tablets.TabletTypes;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 
-public class TabletItemRegister extends BaseItemRegistrator {
+public class InscribingItemRegister extends BaseItemRegistrator {
+
+    public HashMap<SkillItemTier, ScribeInkItem> INK_TIER_MAP = new HashMap<>();
 
     public List<ProtectionTabletItem> ALL_TABLETS = new ArrayList<>();
 
@@ -20,6 +24,12 @@ public class TabletItemRegister extends BaseItemRegistrator {
     public ProtectionTabletItem ANTI_HUNGER = tablet(new ProtectionTabletItem(SkillItemTier.CELESTIAL, TabletTypes.ANTI_HUNGER));
     public ProtectionTabletItem ANTI_GEAR_BREAK = tablet(new ProtectionTabletItem(SkillItemTier.EMPYREAN, TabletTypes.ANTI_GEAR_BREAK));
     public ProtectionTabletItem ANTI_DEATH = tablet(new ProtectionTabletItem(SkillItemTier.DIVINE, TabletTypes.ANTI_DEATH));
+
+    public InscribingItemRegister() {
+        for (SkillItemTier tier : SkillItemTier.values()) {
+            INK_TIER_MAP.put(tier, item(new ScribeInkItem(tier)));
+        }
+    }
 
     ProtectionTabletItem tablet(ProtectionTabletItem c) {
         ALL_TABLETS.add(c);