+++ src/main/java/com/robertx22/age_of_exile/player_skills/enchants/EnchantsEnum.java
@@ -1,0 +1,15 @@
+package com.robertx22.age_of_exile.player_skills.enchants;
+
+public enum EnchantsEnum {
+
+    BONUS_SKILL_DROPS("Bonus Skill Drops", "bonus_skill_drops");
+
+    public String word;
+    public String id;
+
+    EnchantsEnum(String word, String id) {
+        this.word = word;
+        this.id = id;
+    }
+
+}