+++ src/main/java/com/robertx22/age_of_exile/player_skills/items/fishing/ScribeInkItem.java
@@ -1,18 +1,28 @@
-package com.robertx22.age_of_exile.player_skills.items.protection_tablets;
+package com.robertx22.age_of_exile.player_skills.items.fishing;
 
 import com.robertx22.age_of_exile.aoe_data.datapacks.models.IAutoModel;
 import com.robertx22.age_of_exile.aoe_data.datapacks.models.ItemModelManager;
 import com.robertx22.age_of_exile.database.base.CreativeTabs;
+import com.robertx22.age_of_exile.player_skills.items.foods.SkillItemTier;
 import com.robertx22.age_of_exile.uncommon.interfaces.IAutoLocName;
 import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.text.Text;
+import net.minecraft.text.TranslatableText;
 import net.minecraft.util.registry.Registry;
 
-public class BlankTabletItem extends Item implements IAutoLocName, IAutoModel {
-    BlankTabletTier tier;
+public class ScribeInkItem extends Item implements IAutoLocName, IAutoModel {
 
-    public BlankTabletItem(BlankTabletTier tier) {
+    SkillItemTier tier;
+
+    public ScribeInkItem(SkillItemTier tier) {
         super(new Settings().group(CreativeTabs.Inscribing));
         this.tier = tier;
+    }
+
+    @Override
+    public Text getName(ItemStack stack) {
+        return new TranslatableText(this.getTranslationKey()).formatted(tier.format);
     }
 
     @Override
@@ -33,12 +43,12 @@ public class BlankTabletItem extends Item implements IAutoLocName, IAutoModel {
 
     @Override
     public String locNameForLangFile() {
-
-        return tier.prefixName + "Blank Tablet";
+        return tier.word + " Ink";
     }
 
     @Override
     public String GUID() {
-        return "";
+        return "ink/" + tier.tier;
     }
 }
+