+++ src/main/java/com/robertx22/age_of_exile/player_skills/items/inscribing/EnchantmentScrollItem.java
@@ -1,0 +1,134 @@
+package com.robertx22.age_of_exile.player_skills.items.inscribing;
+
+import com.robertx22.age_of_exile.aoe_data.database.player_skills.IsSkillItemUsableUtil;
+import com.robertx22.age_of_exile.aoe_data.datapacks.models.IAutoModel;
+import com.robertx22.age_of_exile.aoe_data.datapacks.models.ItemModelManager;
+import com.robertx22.age_of_exile.aoe_data.datapacks.models.ModelHelper;
+import com.robertx22.age_of_exile.database.base.CreativeTabs;
+import com.robertx22.age_of_exile.database.data.currency.base.IShapelessRecipe;
+import com.robertx22.age_of_exile.mmorpg.ModRegistry;
+import com.robertx22.age_of_exile.player_skills.IReqSkillLevel;
+import com.robertx22.age_of_exile.player_skills.enchants.BaseScribeEnchant;
+import com.robertx22.age_of_exile.player_skills.items.foods.SkillItemTier;
+import com.robertx22.age_of_exile.saveclasses.player_skills.PlayerSkillEnum;
+import com.robertx22.age_of_exile.uncommon.interfaces.IAutoLocName;
+import com.robertx22.age_of_exile.uncommon.utilityclasses.PlayerUtils;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.client.item.TooltipContext;
+import net.minecraft.data.server.recipe.ShapelessRecipeJsonFactory;
+import net.minecraft.enchantment.EnchantmentLevelEntry;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.EnchantedBookItem;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.Items;
+import net.minecraft.text.LiteralText;
+import net.minecraft.text.Text;
+import net.minecraft.text.TranslatableText;
+import net.minecraft.util.Hand;
+import net.minecraft.util.TypedActionResult;
+import net.minecraft.util.registry.Registry;
+import net.minecraft.world.World;
+
+import java.util.List;
+import java.util.function.Supplier;
+
+public class EnchantmentScrollItem extends Item implements IAutoLocName, IAutoModel, IShapelessRecipe, IReqSkillLevel {
+
+    public SkillItemTier tier;
+    public Supplier<BaseScribeEnchant> enchant;
+
+    public EnchantmentScrollItem(Supplier<BaseScribeEnchant> enchant, SkillItemTier tier) {
+        super(new Settings().group(CreativeTabs.Inscribing));
+        this.tier = tier;
+        this.enchant = enchant;
+    }
+
+    @Override
+    public TypedActionResult<ItemStack> use(World world, PlayerEntity user, Hand hand) {
+
+        ItemStack stack = user.getStackInHand(hand);
+        if (!IsSkillItemUsableUtil.canUseItem(user, stack, true)) {
+            return TypedActionResult.fail(stack);
+        }
+
+        stack.decrement(1);
+        ItemStack book = EnchantedBookItem.forEnchantment(new EnchantmentLevelEntry(enchant.get(), tier.tier + 1));
+        PlayerUtils.giveItem(book, user);
+
+        return super.use(world, user, hand);
+
+    }
+
+    @Override
+    public Text getName(ItemStack stack) {
+        return new TranslatableText(this.getTranslationKey()).formatted(tier.format);
+    }
+
+    @Override
+    public AutoLocGroup locNameGroup() {
+        return AutoLocGroup.Enchants;
+    }
+
+    @Override
+    public String locNameLangFileGUID() {
+        return Registry.ITEM.getId(this)
+            .toString();
+    }
+
+    @Override
+    public String locNameForLangFile() {
+        return tier.word + " " + enchant.get()
+            .locNameForLangFile() + " Enchant Scroll";
+    }
+
+    @Override
+    public String GUID() {
+        return "enchant_scroll/" + enchant.get()
+            .GUID() + "/" + tier.tier;
+    }
+
+    @Override
+    public void generateModel(ItemModelManager manager) {
+        // so i dont have to copy icons for each of the 5 tiers
+        ModelHelper helper = new ModelHelper(this, ModelHelper.Type.GENERATED);
+        helper.modelPath = "enchanting_scroll";
+        helper.generate();
+    }
+
+    @Override
+    @Environment(EnvType.CLIENT)
+    public void appendTooltip(ItemStack stack, World world, List<Text> tooltip, TooltipContext context) {
+
+        try {
+
+            tooltip.add(new LiteralText("Tier " + tier.tier + " Inscribing Item."));
+
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+    }
+
+    @Override
+    public ShapelessRecipeJsonFactory getRecipe() {
+        ShapelessRecipeJsonFactory fac = ShapelessRecipeJsonFactory.create(this);
+        fac.input(Items.PAPER);
+        fac.input(this.enchant.get()
+            .getCraftItem()
+            .get());
+        fac.input(ModRegistry.TABLETS.INK_TIER_MAP.get(this.tier));
+        return fac.criterion("player_level", trigger());
+    }
+
+    @Override
+    public PlayerSkillEnum getSkillTypeToCraft() {
+        return PlayerSkillEnum.COOKING;
+    }
+
+    @Override
+    public float getSkillLevelMultiNeeded() {
+        return tier.lvl_req;
+    }
+}